experiment-tracker/experiment-proxy/requirements.txt
fastapi
pydantic
python-multipart
aiohttp
uvicorn
opentelemetry-instrumentation-aiohttp-client

git+https://github.com/xai-demonstrator/xai-demonstrator.git@6f230206fd06fc0b43fc1d430a4ad804d2605123#egg=xaidemo&subdirectory=common/backend-utils
